Update License reference
The theme linked has a custom license and not the OSI approved MIT license that I would expect when the license is stated as the MIT license here. 

I have also ticketed it upstream. https://github.com/PandaSekh/Jekyll-Podcaster/issues/9 however as it stands the license is not the MIT license.